+++ .changeset/nine-baboons-act.md
@@ -2,4 +2,4 @@
 '@finos/legend-application-studio': minor
 ---
 
-Support function overloading ([#1211](https://github.com/finos/legend-studio/issues/1211)). Links to Engine PR ([#759](https://github.com/finos/legend-engine/pull/759)) and SDLC PR ([#515](https://github.com/finos/legend-sdlc/pull/515)).
+Support function overloading ([#1211](https://github.com/finos/legend-studio/issues/1211)).
