+++ .changeset/twenty-glasses-enjoy.md
@@ -2,4 +2,4 @@
 '@finos/legend-graph': minor
 ---
 
-Support function overloading ([#1211](https://github.com/finos/legend-studio/issues/1211)). Links to Engine PR ([#759](https://github.com/finos/legend-engine/pull/759)) and SDLC PR ([#515](https://github.com/finos/legend-sdlc/pull/515)).
+Support function overloading ([#1211](https://github.com/finos/legend-studio/issues/1211)).
