+++ packages/legend-application-studio/src/stores/sidebar-state/LocalChangesState.ts
@@ -193,12 +193,14 @@ export class LocalChangesState {
   openLocalChange(diff: EntityDiff): void {
     const fromEntityGetter = (
       entityPath: string | undefined,
-    ): Entity | undefined =>
-      entityPath
-        ? this.editorStore.changeDetectionState.workspaceLocalLatestRevisionState.entities.find(
-            (e) => e.path === entityPath,
-          )
-        : undefined;
+    ): Entity | undefined => {
+      if (entityPath) {
+        return this.editorStore.changeDetectionState.workspaceLocalLatestRevisionState.entities.find(
+          (e) => e.path === entityPath,
+        );
+      }
+      return undefined;
+    };
     const toEntityGetter = (
       entityPath: string | undefined,
     ): Entity | undefined => {
@@ -249,20 +251,24 @@ export class LocalChangesState {
   openWorkspacePullChange(diff: EntityDiff): void {
     const fromEntityGetter = (
       entityPath: string | undefined,
-    ): Entity | undefined =>
-      entityPath
-        ? this.editorStore.changeDetectionState.workspaceLocalLatestRevisionState.entities.find(
-            (e) => e.path === entityPath,
-          )
-        : undefined;
+    ): Entity | undefined => {
+      if (entityPath) {
+        return this.editorStore.changeDetectionState.workspaceLocalLatestRevisionState.entities.find(
+          (e) => e.path === entityPath,
+        );
+      }
+      return undefined;
+    };
     const toEntityGetter = (
       entityPath: string | undefined,
-    ): Entity | undefined =>
-      entityPath
-        ? this.editorStore.changeDetectionState.workspaceRemoteLatestRevisionState.entities.find(
-            (e) => e.path === entityPath,
-          )
-        : undefined;
+    ): Entity | undefined => {
+      if (entityPath) {
+        return this.editorStore.changeDetectionState.workspaceRemoteLatestRevisionState.entities.find(
+          (e) => e.path === entityPath,
+        );
+      }
+      return undefined;
+    };
     const fromEntity = EntityDiff.shouldOldEntityExist(diff)
       ? guaranteeNonNullable(
           fromEntityGetter(diff.getValidatedOldPath()),
